Refactor rest api to to services.
