+++ .phpstorm.meta.php
@@ -39,7 +39,16 @@ namespace PHPSTORM_META {
 			'tracking'                     => \Google\Web_Stories\Tracking::class,
 			'tinymce'                      => \Google\Web_Stories\TinyMCE::class,
 			'register.widget'              => \Google\Web_Stories\Register_Widget::class,
-			'rest_api_factory'             => \Google\Web_Stories\REST_API_Factory::class,
+			'rest.embed_controller'        => \Google\Web_Stories\REST_API\Embed_Controller::class,
+			'rest.link_controller'         => \Google\Web_Stories\REST_API\Link_Controller::class,
+			'rest.status_check_controller' => \Google\Web_Stories\REST_API\Status_Check_Controller::class,
+			'rest.stories_autosave'        => \Google\Web_Stories\REST_API\Stories_Autosaves_Controller::class,
+			'rest.stories_lock'            => \Google\Web_Stories\REST_API\Stories_Lock_Controller::class,
+			'rest.media'                   => \Google\Web_Stories\REST_API\Stories_Media_Controller::class,
+			'rest.settings'                => \Google\Web_Stories\REST_API\Stories_Settings_Controller::class,
+			'rest.users'                   => \Google\Web_Stories\REST_API\Stories_Users_Controller::class,
+			'rest.template_autosave'       => \Google\Web_Stories\REST_API\Template_Autosaves_Controller::class,
+			'rest.template_lock'           => \Google\Web_Stories\REST_API\Template_Lock_Controller::class,
 			'user_preferences'             => \Google\Web_Stories\User_Preferences::class,
 			'web_stories_block'            => \Google\Web_Stories\Block\Web_Stories_Block::class,
 			'injector'                     => \Google\Web_Stories\Infrastructure\Injector::class,
