+++ includes/REST_API/Embed_Controller.php
@@ -32,7 +32,6 @@ use Google\Web_Stories\Traits\Document_Parser;
 use WP_Error;
 use WP_Http;
 use WP_Post;
-use WP_REST_Controller;
 use WP_REST_Request;
 use WP_REST_Response;
 use WP_REST_Server;
@@ -42,7 +41,7 @@ use WP_REST_Server;
  *
  * Class Embed_Controller
  */
-class Embed_Controller extends WP_REST_Controller {
+class Embed_Controller extends REST_Controller {
 	use Document_Parser;
 	/**
 	 * Constructor.
