+++ includes/REST_API/Stories_Base_Controller.php
@@ -27,6 +27,7 @@
 namespace Google\Web_Stories\REST_API;
 
 use Google\Web_Stories\Decoder;
+use Google\Web_Stories\Services;
 use stdClass;
 use WP_Error;
 use WP_Post;
@@ -59,7 +60,11 @@ class Stories_Base_Controller extends WP_REST_Posts_Controller {
 	public function __construct( $post_type ) {
 		parent::__construct( $post_type );
 		$this->namespace = 'web-stories/v1';
-		$this->decoder   = new Decoder();
+		$injector        = Services::get_injector();
+		if ( ! method_exists( $injector, 'make' ) ) {
+			return;
+		}
+		$this->decoder = $injector->make( Decoder::class );
 	}
 
 	/**
