+++ includes/REST_API/Template_Lock_Controller.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Class Stories_Users_Controller
+ * Class Template_Lock_Controller
  *
  * @package   Google\Web_Stories
  * @copyright 2020 Google LLC
@@ -26,27 +26,19 @@
 
 namespace Google\Web_Stories\REST_API;
 
-use Google\Web_Stories\Infrastructure\Delayed;
-use Google\Web_Stories\Infrastructure\Registerable;
-use Google\Web_Stories\Infrastructure\Service;
-use Google\Web_Stories\Traits\Register_Routes;
-use WP_REST_Users_Controller;
+use Google\Web_Stories\Template_Post_Type;
+use WP_REST_Controller;
 
 /**
- * Stories_Users_Controller class.
+ * Class Template_Lock_Controller
+ *
+ * @package Google\Web_Stories\REST_API
  */
-class Stories_Users_Controller extends WP_REST_Users_Controller implements Service, Delayed, Registerable {
-	use Register_Routes;
+class Template_Lock_Controller extends Lock_Controller {
 	/**
 	 * Constructor.
-	 *
-	 * Override the namespace.
-	 *
-	 * @since 1.2.0
 	 */
 	public function __construct() {
-		parent::__construct();
-		$this->namespace = 'web-stories/v1';
+		parent::__construct( Template_Post_Type::POST_TYPE_SLUG );
 	}
-
 }
